droid/droid/fairo/polymetis/polymetis/python/polymetis/robot_drivers/psyonic_gripper/ability-hand-api/python/psyonic-ability-hand/pyproject.toml
[tool.poetry]
name = "psyonic-ability-hand"
version = "0.1.0"
description = ""
authors = ["Joshua Gleaton <joshua.gleaton@treadlighttech.com>"]

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.6.0"
numpy = "^1.22.3"
smbus2 = "^0.4.1"
rich = "^12.2.0"
pyserial = "^3.5"
evdev = "^1.5.0"
aiofiles = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
